change prepay stage -> payment stage + add enum model builder to dbcontext + modified delete Repo of Doc Template / Cons Task / Project Doc / Transaction

+++ Repository/Implements/ConstructionTaskCategoryRepository.cs
@@ -16,8 +16,11 @@ namespace Repository.Implements
             {
                 using var context = new IdtDbContext();
                 var ctc = context.ConstructionTaskCategories.Where(ctc => ctc.Id == id).FirstOrDefault();
-                context.ConstructionTaskCategories.Remove(ctc);
-                context.SaveChanges();
+                if (ctc != null)
+                {
+                    context.ConstructionTaskCategories.Remove(ctc);
+                    context.SaveChanges();
+                }
             }
             catch
             {
